requirements.txt
pandas
pyarrow
litellm
pydantic
python-dotenv
